move script up a level
