+++ news/125-jlap
@@ -1,0 +1,21 @@
+### Enhancements
+
+* Add experimental `python -m conda_index.json2jlap` script to run after
+  indexing, to create `repodata.jlap` patch sets for incremental repodata
+  downloads. (#125)
+
+### Bug fixes
+
+* <news item>
+
+### Deprecations
+
+* <news item>
+
+### Docs
+
+* <news item>
+
+### Other
+
+* <news itme>
